requirements.txt
django==3.2.13
django-treebeard==4.5.1
pillow==9.1.0
psycopg2-binary==2.9.3
